adjustments from code review
